[n/a] refactor email tempalte to use a new BaseEmail class
