+++ client-mu-plugins/goodbids/src/classes/Plugins/WooCommerce/Emails/AuctionOutbid.php
@@ -10,16 +10,14 @@ namespace GoodBids\Plugins\WooCommerce\Emails;
 
 defined( 'ABSPATH' ) || exit;
 
-use WC_Email;
-use WP_User;
-
+use GoodBids\Plugins\WooCommerce\Emails\BaseEmail;
 /**
- * Auction Outbid extend the custom WooCommerce email class
+ * Auction Outbid extend the custom BaseEmail class
  *
  * @since 1.0.0
- * @extends WC_Email
+ * @extends BaseEmail
  */
-class AuctionOutbid extends WC_Email {
+class AuctionOutbid extends BaseEmail {
 
 	/**
 	 * User ID.
@@ -74,30 +72,10 @@ class AuctionOutbid extends WC_Email {
 	 */
 	public function get_default_subject() {
 		return sprintf(
-			'%s You’ve been outbid',
+			/* translators: %s: site title */
+			__( '[%s] You’ve been outbid', 'goodbids' ),
 			'{site_title}',
 		);
-	}
-
-	/**
-	 * Get the email recipients
-	 *
-	 * @return string
-	 */
-	public function get_recipient(): string {
-		$recipient = parent::get_recipient();
-
-		if ( ! $recipient ) {
-			// this sets the recipient to the settings defined below in init_form_fields()
-			$recipient = $this->get_option( 'recipient' );
-		}
-
-		// if none was entered, just use the WP admin email as a fallback
-		if ( ! $recipient ) {
-			$recipient = get_option( 'admin_email' );
-		}
-
-		return $recipient;
 	}
 
 	/**
@@ -110,18 +88,7 @@ class AuctionOutbid extends WC_Email {
 	public function trigger( $user_id ): void {
 		$this->setup_locale();
 
-		// TODO set up check before sending email
-		if ( $user_id ) {
-			$this->object  = new WP_User( $user_id );
-			$this->user_id = $this->object->ID;
-		}
-
-		if ( ! $this->is_enabled() || ! $this->get_recipient() ) {
-			return;
-		}
-
-		// woohoo, send the email!
-		$this->send( $this->get_recipient(), $this->get_subject(), $this->get_content(), $this->get_headers(), $this->get_attachments() );
+		$this->default_trigger( $user_id );
 
 		$this->restore_locale();
 	}
@@ -136,6 +103,7 @@ class AuctionOutbid extends WC_Email {
 		return wc_get_template_html(
 			$this->template_html,
 			[
+				'instance'      => $this,
 				'email_heading' => $this->get_default_heading(),
 				'button_text'   => $this->get_default_button_text(),
 			]
@@ -153,61 +121,10 @@ class AuctionOutbid extends WC_Email {
 		return wc_get_template_html(
 			$this->template_plain,
 			[
+				'instance'      => $this,
 				'email_heading' => $this->get_default_heading(),
 				'button_text'   => $this->get_default_button_text(),
 			]
 		);
-	}
-
-	/**
-	 * Initialize Settings Form Fields
-	 *
-	 * @since 1.0.0
-	 * @return void
-	 */
-	public function init_form_fields(): void {
-
-		$this->form_fields = [
-			'enabled'            => [
-				'title'   => __( 'Enable/Disable', 'goodbids' ),
-				'type'    => 'checkbox',
-				'label'   => __( 'Enable this email notification', 'goodbids' ),
-				'default' => 'yes',
-			],
-			'subject'            => [
-				'title'       => __( 'Subject', 'goodbids' ),
-				'type'        => 'text',
-				'desc_tip'    => true,
-				'description' => sprintf( 'This controls the email subject line. Leave blank to use the default subject: <code>%s</code>.', $this->get_subject() ),
-				'placeholder' => $this->get_default_subject(),
-				'default'     => '',
-			],
-			'heading'            => [
-				'title'       => __( 'Email Heading', 'goodbids' ),
-				'type'        => 'text',
-				'desc_tip'    => true,
-				'description' => sprintf( __( 'This controls the main heading contained within the email notification. Leave blank to use the default heading: <code>%s</code>.' ), $this->get_heading() ),
-				'placeholder' => $this->get_default_heading(),
-				'default'     => '',
-			],
-			'additional_content' => [
-				'title'       => __( 'Additional content', 'goodbids' ),
-				'description' => __( 'Text to appear below the main email content.', 'goodbids' ),
-				'css'         => 'width:400px; height: 75px;',
-				'placeholder' => __( 'N/A', 'goodbids' ),
-				'type'        => 'textarea',
-				'default'     => $this->get_default_additional_content(),
-				'desc_tip'    => true,
-			],
-			'email_type'         => [
-				'title'       => __( 'Email type', 'goodbids' ),
-				'type'        => 'select',
-				'description' => __( 'Choose which format of email to send.', 'goodbids' ),
-				'default'     => 'html',
-				'class'       => 'email_type wc-enhanced-select',
-				'options'     => $this->get_email_type_options(),
-				'desc_tip'    => true,
-			],
-		];
 	}
 }
