+++ client-mu-plugins/goodbids/src/classes/Plugins/WooCommerce/Emails/AuctionWinnerConfirmation.php
@@ -10,16 +10,15 @@ namespace GoodBids\Plugins\WooCommerce\Emails;
 
 defined( 'ABSPATH' ) || exit;
 
-use WC_Email;
-use WP_User;
+use GoodBids\Plugins\WooCommerce\Emails\BaseEmail;
 
 /**
- * Auction Watchers Live extend the custom WooCommerce email class
+ * Auction Watchers Live extend the custom BaseEmail class
  *
  * @since 1.0.0
- * @extends WC_Email
+ * @extends BaseEmail
  */
-class AuctionWinnerConfirmation extends WC_Email {
+class AuctionWinnerConfirmation extends BaseEmail {
 
 	/**
 	 * User ID.
@@ -54,6 +53,7 @@ class AuctionWinnerConfirmation extends WC_Email {
 	 */
 	public function get_default_subject() {
 		return sprintf(
+			/* translators: %s: site title */
 			__( '[%s] Congratulations, you won!', 'goodbids' ),
 			'{site_title}',
 		);
@@ -67,6 +67,7 @@ class AuctionWinnerConfirmation extends WC_Email {
 	 */
 	public function get_default_heading() {
 		return sprintf(
+			/* translators: %s: reward title */
 			__( 'You generosity earned a %s !', 'goodbids' ),
 			'{auction.rewardTitle}'
 		);
@@ -94,27 +95,6 @@ class AuctionWinnerConfirmation extends WC_Email {
 	}
 
 	/**
-	 * Get the email recipients
-	 *
-	 * @return string
-	 */
-	public function get_recipient(): string {
-		$recipient = parent::get_recipient();
-
-		if ( ! $recipient ) {
-			// this sets the recipient to the settings defined below in init_form_fields()
-			$recipient = $this->get_option( 'recipient' );
-		}
-
-		// if none was entered, just use the WP admin email as a fallback
-		if ( ! $recipient ) {
-			$recipient = get_option( 'admin_email' );
-		}
-
-		return $recipient;
-	}
-
-	/**
 	 * Determine if the email should actually be sent and setup email merge variables
 	 *
 	 * @since 1.0.0
@@ -124,18 +104,7 @@ class AuctionWinnerConfirmation extends WC_Email {
 	public function trigger( $user_id ): void {
 		$this->setup_locale();
 
-		// TODO set up check before sending email
-		if ( $user_id ) {
-			$this->object  = new WP_User( $user_id );
-			$this->user_id = $this->object->ID;
-		}
-
-		if ( ! $this->is_enabled() || ! $this->get_recipient() ) {
-			return;
-		}
-
-		// woohoo, send the email!
-		$this->send( $this->get_recipient(), $this->get_subject(), $this->get_content(), $this->get_headers(), $this->get_attachments() );
+		$this->default_trigger( $user_id );
 
 		$this->restore_locale();
 	}
@@ -175,57 +144,5 @@ class AuctionWinnerConfirmation extends WC_Email {
 				'reward_url'    => $this->get_reward_checkout_url(),
 			]
 		);
-	}
-
-	/**
-	 * Initialize Settings Form Fields
-	 *
-	 * @since 1.0.0
-	 * @return void
-	 */
-	public function init_form_fields(): void {
-
-		$this->form_fields = [
-			'enabled'            => [
-				'title'   => __( 'Enable/Disable', 'goodbids' ),
-				'type'    => 'checkbox',
-				'label'   => __( 'Enable this email notification', 'goodbids' ),
-				'default' => 'yes',
-			],
-			'subject'            => [
-				'title'       => __( 'Subject', 'goodbids' ),
-				'type'        => 'text',
-				'desc_tip'    => true,
-				'description' => sprintf( 'This controls the email subject line. Leave blank to use the default subject: <code>%s</code>.', $this->get_subject() ),
-				'placeholder' => $this->get_default_subject(),
-				'default'     => '',
-			],
-			'heading'            => [
-				'title'       => __( 'Email Heading', 'goodbids' ),
-				'type'        => 'text',
-				'desc_tip'    => true,
-				'description' => sprintf( __( 'This controls the main heading contained within the email notification. Leave blank to use the default heading: <code>%s</code>.' ), $this->get_heading() ),
-				'placeholder' => $this->get_default_heading(),
-				'default'     => '',
-			],
-			'additional_content' => [
-				'title'       => __( 'Additional content', 'goodbids' ),
-				'description' => __( 'Text to appear below the main email content.', 'goodbids' ),
-				'css'         => 'width:400px; height: 75px;',
-				'placeholder' => __( 'N/A', 'goodbids' ),
-				'type'        => 'textarea',
-				'default'     => $this->get_default_additional_content(),
-				'desc_tip'    => true,
-			],
-			'email_type'         => [
-				'title'       => __( 'Email type', 'goodbids' ),
-				'type'        => 'select',
-				'description' => __( 'Choose which format of email to send.', 'goodbids' ),
-				'default'     => 'html',
-				'class'       => 'email_type wc-enhanced-select',
-				'options'     => $this->get_email_type_options(),
-				'desc_tip'    => true,
-			],
-		];
 	}
 }
