+++ client-mu-plugins/goodbids/src/classes/Plugins/WooCommerce/Emails/BaseEmail.php
@@ -1,0 +1,119 @@
+<?php
+/**
+ * BaseEmail: Holds the default settings for all emails
+ *
+ * @since 1.0.0
+ * @package GoodBids
+ */
+
+namespace GoodBids\Plugins\WooCommerce\Emails;
+
+defined( 'ABSPATH' ) || exit;
+
+use WC_Email;
+use WP_User;
+
+/**
+ * BaseEmail: extend the custom BaseEmail class
+ *
+ * @since 1.0.0
+ * @extends WC_Email
+ */
+class BaseEmail extends WC_Email {
+
+
+	/**
+	 * Determine if the email should actually be sent and setup email merge variables
+	 *
+	 * @since 1.0.0
+	 * @param mixed $user_id
+	 * @return void
+	 */
+	public function default_trigger( $user_id ): void {
+		if ( $user_id ) {
+			$this->object  = new WP_User( $user_id );
+			$this->user_id = $this->object->ID;
+		}
+
+		if ( ! $this->is_enabled() || ! $this->get_recipient() ) {
+			return;
+		}
+
+		// woohoo, send the email!
+		$this->send( $this->get_recipient(), $this->get_subject(), $this->get_content(), $this->get_headers(), $this->get_attachments() );
+	}
+
+
+	/**
+	 * Get the default email recipients
+	 *
+	 * @return string
+	 */
+	public function get_recipient(): string {
+		$recipient = parent::get_recipient();
+
+		if ( ! $recipient ) {
+			// this sets the recipient to the settings defined below in init_form_fields()
+			$recipient = $this->get_option( 'recipient' );
+		}
+
+		// if none was entered, just use the WP admin email as a fallback
+		if ( ! $recipient ) {
+			$recipient = get_option( 'admin_email' );
+		}
+
+		return $recipient;
+	}
+
+	/**
+	 * Initialize Settings Form Fields
+	 *
+	 * @since 1.0.0
+	 * @return void
+	 */
+	public function init_form_fields(): void {
+
+		$this->form_fields = [
+			'enabled'            => [
+				'title'   => __( 'Enable/Disable', 'goodbids' ),
+				'type'    => 'checkbox',
+				'label'   => __( 'Enable this email notification', 'goodbids' ),
+				'default' => 'yes',
+			],
+			'subject'            => [
+				'title'       => __( 'Subject', 'goodbids' ),
+				'type'        => 'text',
+				'desc_tip'    => true,
+				'description' => sprintf( 'This controls the email subject line. Leave blank to use the default subject: <code>%s</code>.', $this->get_subject() ),
+				'placeholder' => $this->get_default_subject(),
+				'default'     => '',
+			],
+			'heading'            => [
+				'title'       => __( 'Email Heading', 'goodbids' ),
+				'type'        => 'text',
+				'desc_tip'    => true,
+				'description' => sprintf( __( 'This controls the main heading contained within the email notification. Leave blank to use the default heading: <code>%s</code>.' ), $this->get_heading() ),
+				'placeholder' => $this->get_default_heading(),
+				'default'     => '',
+			],
+			'additional_content' => [
+				'title'       => __( 'Additional content', 'goodbids' ),
+				'description' => __( 'Text to appear below the main email content.', 'goodbids' ),
+				'css'         => 'width:400px; height: 75px;',
+				'placeholder' => __( 'N/A', 'goodbids' ),
+				'type'        => 'textarea',
+				'default'     => $this->get_default_additional_content(),
+				'desc_tip'    => true,
+			],
+			'email_type'         => [
+				'title'       => __( 'Email type', 'goodbids' ),
+				'type'        => 'select',
+				'description' => __( 'Choose which format of email to send.', 'goodbids' ),
+				'default'     => 'html',
+				'class'       => 'email_type wc-enhanced-select',
+				'options'     => $this->get_email_type_options(),
+				'desc_tip'    => true,
+			],
+		];
+	}
+}
